+++ sismic/code/dummy.py
@@ -16,8 +16,8 @@ class DummyEvaluator(Evaluator):
     def context(self):
         return dict()
 
-    def _evaluate_code(self, code: str, additional_context: Mapping=None) -> bool:
+    def _evaluate_code(self, code: str, *, additional_context: Mapping=None) -> bool:
         return True
 
-    def _execute_code(self, code: str, additional_context: Mapping=None):
+    def _execute_code(self, code: str, *, additional_context: Mapping=None) -> None:
         return